requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
pyobjc-framework-Accessibility>=10.0
pyobjc-framework-ApplicationServices>=10.0
websocket-client>=1.6.0
python-dotenv>=1.0.0

